+++ x/ibc/04-channel/keeper/timeout_test.go
@@ -3,10 +3,9 @@ package keeper_test
 import (
 	"fmt"
 
-	"github.com/cosmos/cosmos-sdk/x/capability"
+	capabilitytypes "github.com/cosmos/cosmos-sdk/x/capability/types"
 	connection "github.com/cosmos/cosmos-sdk/x/ibc/03-connection"
 	"github.com/cosmos/cosmos-sdk/x/ibc/04-channel/types"
-	commitmentexported "github.com/cosmos/cosmos-sdk/x/ibc/23-commitment/exported"
 	host "github.com/cosmos/cosmos-sdk/x/ibc/24-host"
 )
 
@@ -112,7 +111,7 @@ func (suite *KeeperTestSuite) TestTimeoutExecuted() {
 
 	var (
 		packet  types.Packet
-		chanCap *capability.Capability
+		chanCap *capabilitytypes.Capability
 	)
 
 	testCases := []testCase{
@@ -125,7 +124,7 @@ func (suite *KeeperTestSuite) TestTimeoutExecuted() {
 		{"incorrect capability", func() {
 			packet = types.NewPacket(newMockTimeoutPacket().GetBytes(), 1, testPort1, testChannel1, testPort2, testChannel2, timeoutHeight, disabledTimeoutTimestamp)
 			suite.chainA.createChannel(testPort1, testChannel1, testPort2, testChannel2, types.OPEN, types.ORDERED, testConnectionIDA)
-			chanCap = capability.NewCapability(100)
+			chanCap = capabilitytypes.NewCapability(100)
 		}, false},
 	}
 
@@ -242,7 +241,7 @@ func (suite *KeeperTestSuite) TestTimeoutOnClose() {
 	for i, tc := range testCases {
 		tc := tc
 		suite.Run(fmt.Sprintf("Case %s, %d/%d tests", tc.msg, i, len(testCases)), func() {
-			var proof commitmentexported.Proof
+			var proof []byte
 
 			suite.SetupTest() // reset
 			tc.malleate()