+++ x/ibc/07-tendermint/types/client_state.go
@@ -3,6 +3,7 @@ package types
 import (
 	"time"
 
+	ics23 "github.com/confio/ics23/go"
 	tmmath "github.com/tendermint/tendermint/libs/math"
 	lite "github.com/tendermint/tendermint/lite2"
 
@@ -46,6 +47,8 @@ type ClientState struct {
 
 	// Last Header that was stored by client
 	LastHeader Header `json:"last_header" yaml:"last_header"`
+
+	ProofSpecs []*ics23.ProofSpec `json:"proof_specs" yaml:"proof_specs"`
 }
 
 // InitializeFromMsg creates a tendermint client state from a CreateClientMsg
@@ -53,7 +56,7 @@ func InitializeFromMsg(msg MsgCreateClient) (ClientState, error) {
 	return Initialize(
 		msg.GetClientID(), msg.TrustLevel,
 		msg.TrustingPeriod, msg.UnbondingPeriod, msg.MaxClockDrift,
-		msg.Header,
+		msg.Header, msg.ProofSpecs,
 	)
 }
 
@@ -62,17 +65,10 @@ func InitializeFromMsg(msg MsgCreateClient) (ClientState, error) {
 func Initialize(
 	id string, trustLevel tmmath.Fraction,
 	trustingPeriod, ubdPeriod, maxClockDrift time.Duration,
-	header Header,
+	header Header, specs []*ics23.ProofSpec,
 ) (ClientState, error) {
+	clientState := NewClientState(id, trustLevel, trustingPeriod, ubdPeriod, maxClockDrift, header, specs)
 
-	if trustingPeriod >= ubdPeriod {
-		return ClientState{}, sdkerrors.Wrapf(
-			ErrInvalidTrustingPeriod,
-			"trusting period (%s) should be < unbonding period (%s)", trustingPeriod, ubdPeriod,
-		)
-	}
-
-	clientState := NewClientState(id, trustLevel, trustingPeriod, ubdPeriod, maxClockDrift, header)
 	return clientState, nil
 }
 
@@ -80,7 +76,7 @@ func Initialize(
 func NewClientState(
 	id string, trustLevel tmmath.Fraction,
 	trustingPeriod, ubdPeriod, maxClockDrift time.Duration,
-	header Header,
+	header Header, specs []*ics23.ProofSpec,
 ) ClientState {
 	return ClientState{
 		ID:              id,
@@ -90,6 +86,7 @@ func NewClientState(
 		MaxClockDrift:   maxClockDrift,
 		LastHeader:      header,
 		FrozenHeight:    0,
+		ProofSpecs:      specs,
 	}
 }
 
@@ -143,23 +140,47 @@ func (cs ClientState) Validate() error {
 	if cs.MaxClockDrift == 0 {
 		return sdkerrors.Wrap(ErrInvalidMaxClockDrift, "max clock drift cannot be zero")
 	}
+	if cs.TrustingPeriod >= cs.UnbondingPeriod {
+		return sdkerrors.Wrapf(
+			ErrInvalidTrustingPeriod,
+			"trusting period (%s) should be < unbonding period (%s)", cs.TrustingPeriod, cs.UnbondingPeriod,
+		)
+	}
+	// Validate ProofSpecs
+	if cs.ProofSpecs == nil {
+		return sdkerrors.Wrap(ErrInvalidProofSpecs, "proof specs cannot be nil for tm client")
+	}
+	for _, spec := range cs.ProofSpecs {
+		if spec == nil {
+			return sdkerrors.Wrap(ErrInvalidProofSpecs, "proof spec cannot be nil")
+		}
+	}
+
 	return cs.LastHeader.ValidateBasic(cs.GetChainID())
+}
+
+// GetProofSpecs returns the format the client expects for proof verification
+// as a string array specifying the proof type for each position in chained proof
+func (cs ClientState) GetProofSpecs() []*ics23.ProofSpec {
+	return cs.ProofSpecs
 }
 
 // VerifyClientConsensusState verifies a proof of the consensus state of the
 // Tendermint client stored on the target machine.
 func (cs ClientState) VerifyClientConsensusState(
 	_ sdk.KVStore,
-	cdc *codec.Codec,
+	cdc codec.Marshaler,
+	aminoCdc *codec.Codec,
 	provingRoot commitmentexported.Root,
 	height uint64,
 	counterpartyClientIdentifier string,
 	consensusHeight uint64,
 	prefix commitmentexported.Prefix,
-	proof commitmentexported.Proof,
+	proof []byte,
 	consensusState clientexported.ConsensusState,
 ) error {
-	if err := validateVerificationArgs(cs, height, prefix, proof, consensusState); err != nil {
+	merkleProof, err := sanitizeVerificationArgs(cdc, cs, height, prefix, proof, consensusState)
+	if err != nil {
 		return err
 	}
 
@@ -169,12 +190,12 @@ func (cs ClientState) VerifyClientConsensusState(
 		return err
 	}
 
-	bz, err := cdc.MarshalBinaryBare(consensusState)
+	bz, err := aminoCdc.MarshalBinaryBare(consensusState)
 	if err != nil {
 		return err
 	}
 
-	if err := proof.VerifyMembership(provingRoot, path, bz); err != nil {
+	if err := merkleProof.VerifyMembership(cs.ProofSpecs, provingRoot, path, bz); err != nil {
 		return sdkerrors.Wrap(clienttypes.ErrFailedClientConsensusStateVerification, err.Error())
 	}
 
@@ -188,12 +209,13 @@ func (cs ClientState) VerifyConnectionState(
 	cdc codec.Marshaler,
 	height uint64,
 	prefix commitmentexported.Prefix,
-	proof commitmentexported.Proof,
+	proof []byte,
 	connectionID string,
 	connectionEnd connectionexported.ConnectionI,
 	consensusState clientexported.ConsensusState,
 ) error {
-	if err := validateVerificationArgs(cs, height, prefix, proof, consensusState); err != nil {
+	merkleProof, err := sanitizeVerificationArgs(cdc, cs, height, prefix, proof, consensusState)
+	if err != nil {
 		return err
 	}
 
@@ -212,7 +234,7 @@ func (cs ClientState) VerifyConnectionState(
 		return err
 	}
 
-	if err := proof.VerifyMembership(consensusState.GetRoot(), path, bz); err != nil {
+	if err := merkleProof.VerifyMembership(cs.ProofSpecs, consensusState.GetRoot(), path, bz); err != nil {
 		return sdkerrors.Wrap(clienttypes.ErrFailedConnectionStateVerification, err.Error())
 	}
 
@@ -226,13 +248,14 @@ func (cs ClientState) VerifyChannelState(
 	cdc codec.Marshaler,
 	height uint64,
 	prefix commitmentexported.Prefix,
-	proof commitmentexported.Proof,
+	proof []byte,
 	portID,
 	channelID string,
 	channel channelexported.ChannelI,
 	consensusState clientexported.ConsensusState,
 ) error {
-	if err := validateVerificationArgs(cs, height, prefix, proof, consensusState); err != nil {
+	merkleProof, err := sanitizeVerificationArgs(cdc, cs, height, prefix, proof, consensusState)
+	if err != nil {
 		return err
 	}
 
@@ -251,7 +274,7 @@ func (cs ClientState) VerifyChannelState(
 		return err
 	}
 
-	if err := proof.VerifyMembership(consensusState.GetRoot(), path, bz); err != nil {
+	if err := merkleProof.VerifyMembership(cs.ProofSpecs, consensusState.GetRoot(), path, bz); err != nil {
 		return sdkerrors.Wrap(clienttypes.ErrFailedChannelStateVerification, err.Error())
 	}
 
@@ -262,16 +285,18 @@ func (cs ClientState) VerifyChannelState(
 // the specified port, specified channel, and specified sequence.
 func (cs ClientState) VerifyPacketCommitment(
 	_ sdk.KVStore,
+	cdc codec.Marshaler,
 	height uint64,
 	prefix commitmentexported.Prefix,
-	proof commitmentexported.Proof,
+	proof []byte,
 	portID,
 	channelID string,
 	sequence uint64,
 	commitmentBytes []byte,
 	consensusState clientexported.ConsensusState,
 ) error {
-	if err := validateVerificationArgs(cs, height, prefix, proof, consensusState); err != nil {
+	merkleProof, err := sanitizeVerificationArgs(cdc, cs, height, prefix, proof, consensusState)
+	if err != nil {
 		return err
 	}
 
@@ -280,7 +305,7 @@ func (cs ClientState) VerifyPacketCommitment(
 		return err
 	}
 
-	if err := proof.VerifyMembership(consensusState.GetRoot(), path, commitmentBytes); err != nil {
+	if err := merkleProof.VerifyMembership(cs.ProofSpecs, consensusState.GetRoot(), path, commitmentBytes); err != nil {
 		return sdkerrors.Wrap(clienttypes.ErrFailedPacketCommitmentVerification, err.Error())
 	}
 
@@ -291,16 +316,18 @@ func (cs ClientState) VerifyPacketCommitment(
 // acknowledgement at the specified port, specified channel, and specified sequence.
 func (cs ClientState) VerifyPacketAcknowledgement(
 	_ sdk.KVStore,
+	cdc codec.Marshaler,
 	height uint64,
 	prefix commitmentexported.Prefix,
-	proof commitmentexported.Proof,
+	proof []byte,
 	portID,
 	channelID string,
 	sequence uint64,
 	acknowledgement []byte,
 	consensusState clientexported.ConsensusState,
 ) error {
-	if err := validateVerificationArgs(cs, height, prefix, proof, consensusState); err != nil {
+	merkleProof, err := sanitizeVerificationArgs(cdc, cs, height, prefix, proof, consensusState)
+	if err != nil {
 		return err
 	}
 
@@ -309,7 +336,7 @@ func (cs ClientState) VerifyPacketAcknowledgement(
 		return err
 	}
 
-	if err := proof.VerifyMembership(consensusState.GetRoot(), path, channeltypes.CommitAcknowledgement(acknowledgement)); err != nil {
+	if err := merkleProof.VerifyMembership(cs.ProofSpecs, consensusState.GetRoot(), path, channeltypes.CommitAcknowledgement(acknowledgement)); err != nil {
 		return sdkerrors.Wrap(clienttypes.ErrFailedPacketAckVerification, err.Error())
 	}
 
@@ -321,15 +348,17 @@ func (cs ClientState) VerifyPacketAcknowledgement(
 // specified sequence.
 func (cs ClientState) VerifyPacketAcknowledgementAbsence(
 	_ sdk.KVStore,
+	cdc codec.Marshaler,
 	height uint64,
 	prefix commitmentexported.Prefix,
-	proof commitmentexported.Proof,
+	proof []byte,
 	portID,
 	channelID string,
 	sequence uint64,
 	consensusState clientexported.ConsensusState,
 ) error {
-	if err := validateVerificationArgs(cs, height, prefix, proof, consensusState); err != nil {
+	merkleProof, err := sanitizeVerificationArgs(cdc, cs, height, prefix, proof, consensusState)
+	if err != nil {
 		return err
 	}
 
@@ -338,7 +367,7 @@ func (cs ClientState) VerifyPacketAcknowledgementAbsence(
 		return err
 	}
 
-	if err := proof.VerifyNonMembership(consensusState.GetRoot(), path); err != nil {
+	if err := merkleProof.VerifyNonMembership(cs.ProofSpecs, consensusState.GetRoot(), path); err != nil {
 		return sdkerrors.Wrap(clienttypes.ErrFailedPacketAckAbsenceVerification, err.Error())
 	}
 
@@ -349,15 +378,17 @@ func (cs ClientState) VerifyPacketAcknowledgementAbsence(
 // received of the specified channel at the specified port.
 func (cs ClientState) VerifyNextSequenceRecv(
 	_ sdk.KVStore,
+	cdc codec.Marshaler,
 	height uint64,
 	prefix commitmentexported.Prefix,
-	proof commitmentexported.Proof,
+	proof []byte,
 	portID,
 	channelID string,
 	nextSequenceRecv uint64,
 	consensusState clientexported.ConsensusState,
 ) error {
-	if err := validateVerificationArgs(cs, height, prefix, proof, consensusState); err != nil {
+	merkleProof, err := sanitizeVerificationArgs(cdc, cs, height, prefix, proof, consensusState)
+	if err != nil {
 		return err
 	}
 
@@ -368,59 +399,60 @@ func (cs ClientState) VerifyNextSequenceRecv(
 
 	bz := sdk.Uint64ToBigEndian(nextSequenceRecv)
 
-	if err := proof.VerifyMembership(consensusState.GetRoot(), path, bz); err != nil {
+	if err := merkleProof.VerifyMembership(cs.ProofSpecs, consensusState.GetRoot(), path, bz); err != nil {
 		return sdkerrors.Wrap(clienttypes.ErrFailedNextSeqRecvVerification, err.Error())
 	}
 
 	return nil
 }
 
-// validateVerificationArgs perfoms the basic checks on the arguments that are
-// shared between the verification functions.
-func validateVerificationArgs(
+// sanitizeVerificationArgs perfoms the basic checks on the arguments that are
+// shared between the verification functions and returns the unmarshalled
+// merkle proof and an error if one occurred.
+func sanitizeVerificationArgs(
+	cdc codec.Marshaler,
 	cs ClientState,
 	height uint64,
 	prefix commitmentexported.Prefix,
-	proof commitmentexported.Proof,
+	proof []byte,
 	consensusState clientexported.ConsensusState,
-) error {
+) (merkleProof commitmenttypes.MerkleProof, err error) {
 	if cs.GetLatestHeight() < height {
-		return sdkerrors.Wrapf(
+		return commitmenttypes.MerkleProof{}, sdkerrors.Wrapf(
 			sdkerrors.ErrInvalidHeight,
 			"client state (%s) height < proof height (%d < %d)", cs.ID, cs.GetLatestHeight(), height,
 		)
 	}
 
 	if cs.IsFrozen() && cs.FrozenHeight <= height {
-		return clienttypes.ErrClientFrozen
+		return commitmenttypes.MerkleProof{}, clienttypes.ErrClientFrozen
 	}
 
 	if prefix == nil {
-		return sdkerrors.Wrap(commitmenttypes.ErrInvalidPrefix, "prefix cannot be empty")
+		return commitmenttypes.MerkleProof{}, sdkerrors.Wrap(commitmenttypes.ErrInvalidPrefix, "prefix cannot be empty")
 	}
 
 	_, ok := prefix.(*commitmenttypes.MerklePrefix)
 	if !ok {
-		return sdkerrors.Wrapf(commitmenttypes.ErrInvalidPrefix, "invalid prefix type %T, expected *MerklePrefix", prefix)
+		return commitmenttypes.MerkleProof{}, sdkerrors.Wrapf(commitmenttypes.ErrInvalidPrefix, "invalid prefix type %T, expected *MerklePrefix", prefix)
 	}
 
 	if proof == nil {
-		return sdkerrors.Wrap(commitmenttypes.ErrInvalidProof, "proof cannot be empty")
+		return commitmenttypes.MerkleProof{}, sdkerrors.Wrap(commitmenttypes.ErrInvalidProof, "proof cannot be empty")
 	}
 
-	_, ok = proof.(commitmenttypes.MerkleProof)
-	if !ok {
-		return sdkerrors.Wrapf(commitmenttypes.ErrInvalidProof, "invalid proof type %T, expected MerkleProof", proof)
+	if err = cdc.UnmarshalBinaryBare(proof, &merkleProof); err != nil {
+		return commitmenttypes.MerkleProof{}, sdkerrors.Wrap(commitmenttypes.ErrInvalidProof, "failed to unmarshal proof into commitment merkle proof")
 	}
 
 	if consensusState == nil {
-		return sdkerrors.Wrap(clienttypes.ErrInvalidConsensus, "consensus state cannot be empty")
+		return commitmenttypes.MerkleProof{}, sdkerrors.Wrap(clienttypes.ErrInvalidConsensus, "consensus state cannot be empty")
 	}
 
 	_, ok = consensusState.(ConsensusState)
 	if !ok {
-		return sdkerrors.Wrapf(clienttypes.ErrInvalidConsensus, "invalid consensus type %T, expected %T", consensusState, ConsensusState{})
+		return commitmenttypes.MerkleProof{}, sdkerrors.Wrapf(clienttypes.ErrInvalidConsensus, "invalid consensus type %T, expected %T", consensusState, ConsensusState{})
 	}
 
-	return nil
+	return merkleProof, nil
 }